Revert "Merge pull request #2129 from ceph/wip-librbd-oc"
This reverts commit 1482716e6b4d727f4bd5d82df5e0073972f25f58, reversing
changes made to 89368f2c7d4defd8da0afb5ea9c3e426a7831ab5.

The dirty_or_tx list is used by flush_set, which means we can
resubmit new IOs for writes that are already in progress.  This
has a compounding effect that overwhelms the OSDs with dup IOs
and stalls out the client.

See, for example, teh failues in this run:
  /a/sage-2014-07-25_17:14:20-fs-wip-msgr-testing-basic-plana

The fix is probably pretty simple, but reverting for now to make
the tests pass.

Signed-off-by: Sage Weil <sage@inktank.com>