test/cli-integration/rbd: fix trailing space
Newer versions of json.tool remove the trailing ' ' after the comma.  Add
it back in with sed so that the .t works on both old and new versions, and
so that we don't have to remove the trailing spaces from all of the test
cases.

Backport: firefly
Fixes: #8920
Signed-off-by: Sage Weil <sage@redhat.com>
(cherry picked from commit 605064dc685aa25cc7d58ec18b6449a3ce476d01)

Conflicts:
	src/test/cli-integration/rbd/defaults.t