rgw: clean up some locator use
We don't use locators anymore. New objects are always created having the
locator equals to the object name. Don't allow this to happen through
magic, just get rid of code that is setting the locator. Legacy objects
should continue to work as we'll still decode the locator, and set it on
the ioctx where needed.

Signed-off-by: Yehuda Sadeh <yehuda@redhat.com>