rgw_users: Generate an id(access_key) during SWIFT key creation
Description: An blank/empty id(access_key) is generated during SWIFT key creation
and is entered into the .users.swift pool; instead of the specified
subuser(user:subuser). "radosgw-admin user info" will show the correct
information; but the same subuser will not be listed under .users.swift
pool. This caused rgw to return with an 403 Forbidden error for the
subuser.

Fixes: #9155

Signed-off-by: Dhiraj Kamble <dhiraj.kamble@sandisk.com>