use 'mon create-initial' always
But don't error if it fails, as this would mean that the monitors
are just taking longer to form quorum. Go and try the next block which will
wait up to 15 minutes for a successful gatherkeys to happen (that only works
if monitors have formed quorum).

Signed-off-by: Alfredo Deza <alfredo.deza@inktank.com>
(cherry picked from commit 70a1f18adfdf673b983cc7f5f2c78abb899001c7)