mon: pool create must not always create a ruleset
The implicit creation of a ruleset when creating a pool is convenient
when nothing is specified. However, if the caller sets a ruleset name,
it should not implicitly create it but return ENOENT instead. Silently
creating a ruleset when there is a typo in the ruleset name is
confusing.

http://tracker.ceph.com/issues/9304 Fixes: #9304

Signed-off-by: Loic Dachary <loic-201408@dachary.org>
(cherry picked from commit 8b27997be90eddad930da30b38d6ad22d4621635)

Conflicts:
	src/mon/OSDMonitor.cc