osd: initialize purged_snap on backfill start; restart backfill if change
If we backfill a PG to a new OSD, we currently neglect to initialize
purged_snaps.  As a result, the first time the snaptrimmer runs it has to
churn through every deleted snap for all time, and to make matters worse
does so in one go with the PG lock held.  This leads to badness on any
cluster with a significant number of removed snaps that experiences
backfill.

Resolve this by initializing purged_snaps when we begin backfill.  The
backfill itself will clear out any stray snaps and ensure the object set
is in sync with purged_snaps.  Note that purged_snaps on the primary
that is driving backfill will not change during this period as the
snaptrimmer is not scheduled unless the PG is clean (which it won't be
during backfill).

Restart backfill if we have a last_backfill != max and a purged_snaps that
does not match ours, so we can ensure that it is brought into sync.  This
will generally never happen unless a partially backfilled OSD is down for
long enough that the PG becomes fully clean and purges some snaps.  In that
case, it most likely needs to undergo backfill anyway.

Fixes: #9487
Backfill: firefly, dumpling
Signed-off-by: Sage Weil <sage@redhat.com>
(cherry picked from commit bd5a90ccc17ba55236476907f602f332f8d484bc)

Conflicts:

	src/osd/PG.cc