mon: MDSMonitor: check all conditions are met *before* osdmon proposal
We should not allow ourselves to request the osdmon to propose before we
know for sure that we meet the required conditions to go through with
our own state change.  Even if we still can't guarantee that our
proposal is going to be committed, we shouldn't just change the osdmon's
state just because we can.  This way, at least, we make sure that our
checks hold up before doing anything with side-effects.

Fixes: #10643

Signed-off-by: Joao Eduardo Luis <joao@redhat.com>
(cherry picked from commit 1d1ae413e9f0aafa314af903bcf6b815f6558082)